# mail-worker/openapi/cloudmail-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudMail API",
    "version": "1.0.0",
    "description": "Versioned CloudMail API contract. /api remains a compatibility alias for Cloud-Mail 3.0.0 clients; /api/v1 is the stable v1 namespace."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Stable v1 API"
    },
    {
      "url": "/api",
      "description": "Legacy compatibility alias"
    }
  ],
  "paths": {
...
    "/login": {
      "post": {
        "operationId": "login_post_login",
        "summary": "POST /login",
        "responses": {
          "200": {
            "description": "Successful CloudMail response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Permission denied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/logout": {
      "delete": {
        "operationId": "login_delete_logout",
        "summary": "DELETE /logout",
        "responses": {
          "200": {
            "description": "Successful CloudMail response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Permission denied"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "webSession": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/refresh": {
      "post": {
        "operationId": "login_post_refresh",
        "summary": "POST /refresh",
        "responses": {
          "200": {
            "description": "Successful CloudMail response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Permission denied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/register": {
      "post": {
        "operationId": "login_post_register",
        "summary": "POST /register",
        "responses": {
          "200": {
            "description": "Successful CloudMail response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Permission denied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "webSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-cfmail_session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}